condo_system_project/Backend/login_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
sqlalchemy
alembic
requests
